Add Gemfile and puppet 4 checks
In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. This patch also
corrects lint errors caught by the puppet-lint-empty-check gem.

The $vhost_extra default was changed from empty string to undef, which
will be evaluated as nil in the template and will result in the same
behavior (printing nothing) if left unset.

Change-Id: I31588fb2d4c1894ad9785cd001100a83f4736f2f