Convert httpd::vhost to apache::vhost::custom
For the most part, we can use the httpd::vhost defined type with
templates defined in other modules as long as the variables in the
templates are looked up with the scope object. This does not work at
all if the httpd::vhost is declared within another defined type as
opposed to a class. This patch converts the module to use the upstream
apache::vhost::custom defined type. The benefit of this approach is
instead of passing the name of a template file which will be evaluated
within httpd::vhost and not have access to any of the module's
variables, we evaluate the template within the module and pass the
content to the defined type.

The httpd::vhost defined type managed firewall rules while the upstream
version does not, so add that in too.

Also switch to the upstream apache class, and remove the httpd::mod
resource that we don't need to support any more.

Change-Id: I3a45e99e3355d1e59d57d3d989cc03d253529bf2