Use httpd::vhost $content parameter
Use the new $content parameter of httpd::vhost instead of $template.
This way, the template gets rendered within the scope where it assumes
its variables are, and doesn't need to use the scope object which
doesn't work within a defined type. This will ensure the template keeps
working on puppet 4.

Depends-On: https://review.openstack.org/570824

Change-Id: I3a45e99e3355d1e59d57d3d989cc03d253529bf2