Convert httpd::vhost to apache::vhost::custom
For the most part, we can use the httpd::vhost defined type with
templates defined in other modules as long as the variables in the
templates are looked up with the scope object. This does not work at
all if the httpd::vhost is declared within another defined type as
opposed to a class, which means without this patch the eavesdrop vhost
template won't work with puppet 4. This patch starts transitioning the
meetbot::site resource to use the upstream apache::vhost::custom
resource depending on which apache module is in effect. The benefit of
using the upstream resource is instead of passing the name of a template
file which will be evaluated within httpd::vhost and not have access to
any of the module's variables, we evaluate the template within the
module and pass the content to the vhost resource.

Also clean up the httpd::mod resource that we don't need to support any
more, and remove the explicit httpd class declaration since we do that
in system-config and want to transition to using the other one.

Change-Id: I3a45e99e3355d1e59d57d3d989cc03d253529bf2