Use httpd::vhost $content parameter
Use the new $content parameter of httpd::vhost instead of $template.
This way, the template gets rendered within the scope where it assumes
most of its variables are, and doesn't need to use the scope object
which doesn't work within a defined type. This will ensure the template
keeps working on puppet 4. This also requires adjusting some of the
template parameters that had previously been correctly scoped from the
httpd::vhost defined type.

Depends-On: https://review.openstack.org/570824

Change-Id: I3a45e99e3355d1e59d57d3d989cc03d253529bf2